Moved the default value generation to the app component. So a proper loading widget can be displayed.
